UI button model refactor & auth extension
