+++ lib/widgets/dialogs/dialog.dart
@@ -1,17 +1,10 @@
 import 'package:flutter/material.dart';
-import 'package:fokus/model/button_type.dart';
+import 'package:fokus/model/ui/button.dart';
 import 'package:fokus/utils/app_locales.dart';
-
-class DialogButton {
-	final ButtonType type;
-	final void Function() action;
-
-	DialogButton(this.type, this.action);
-}
 
 class AppDialog extends StatelessWidget {
 	final String titleKey, textKey;
-	final List<DialogButton> buttons;
+	final List<UIButton> buttons;
 
 	AppDialog({@required this.titleKey, this.textKey, this.buttons = const []});
 
@@ -22,7 +15,7 @@ class AppDialog extends StatelessWidget {
 			content: Text(AppLocales.of(context).translate(textKey)),
 			actions: buttons.map((button) => FlatButton(
 				child: Text(AppLocales.of(context).translate(button.type.key)),
-				onPressed: () => button.action(),
+				onPressed: () => button.action(context),
 			)).toList(),
 		);
 	}
